+++ jacop-core/src/main/java/org/jacop/core/IntVar.java
@@ -119,7 +119,7 @@ public class IntVar extends Var {
    */
   public IntVar(Store store, String name, int min, int max) {
 
-    if ((long) max - (long) min > 63L) {
+    if ((long) max - min > 63L) {
       commonInitialization(store, name, new IntervalDomain(min, max));
     } else {
       commonInitialization(store, name, new SmallDenseDomain(min, max));
