+++ jacop-floats/src/main/java/org/jacop/floats/constraints/XeqP.java
@@ -81,14 +81,14 @@ public class XeqP extends Constraint implements SatisfiedPresent {
 
       // domain consistency
       int xMin;
-      if (Math.abs(p.min()) < (double) IntDomain.MAX_INT) {
+      if (Math.abs(p.min()) < IntDomain.MAX_INT) {
         xMin = (int) (Math.round(Math.ceil(p.min())));
       } else {
         xMin = IntDomain.MIN_INT;
       }
 
       int xMax;
-      if (Math.abs(p.max()) < (double) IntDomain.MAX_INT) {
+      if (Math.abs(p.max()) < IntDomain.MAX_INT) {
         xMax = (int) (Math.round(Math.floor(p.max())));
       } else {
         xMax = IntDomain.MAX_INT;
